requirements.txt
accelerate
datasets==2.13.1
sentencepiece!=0.1.92
transformers<=4.33.3
neptune
pdbpp
notebook
protobuf==3.20.*
pyyaml
pynvml
hydra-core
evaluate
nltk
absl-py
rouge_score
deepchem
selfies
tabulate
schedule
scikit-learn
python-Levenshtein
fsspec==2023.6.0